Count only 3 in a row pings and make API more stateful
Which means that we will ignore pings that do not have
the same result three times in a row.

And API now synchronizes the downtimes between DB and Kuma.
It won't only look at the last ping, and then say that
the node is down. Instead, it will parse all 50 pings
returned by Kuma.